__tests__/board.test.ts
// __tests__/board.test.ts


import { Board, Tile } from '../src/board';




describe('Board', () => {
    let board: Board<Tile>; // 使用你的Tile枚举类型
    const rows = 8;
    const cols = 8;
    const mockGenerator = jest.fn(); // 创建一个模拟的生成器函数

    beforeEach(() => {
        // 在每个测试用例开始前重置模拟生成器并创建新的Board实例
        mockGenerator.mockReset().mockReturnValue(Tile.A); // 默认生成A类型的瓷砖
        board = new Board(rows, cols, mockGenerator);
    });

    test('should create a board with correct rows and columns', () => {
        expect(board).toBeDefined();
        expect(board.rows).toBe(rows); // 假设你的Board类有一个rows属性
        expect(board.cols).toBe(cols); // 假设你的Board类有一个cols属性
    });

    test('canMove should validate moves correctly', () => {
        // 测试canMove逻辑，确保它在合法和非法移动上返回正确的结果
        // 你需要根据canMove方法的实现来编写具体的测试用例
    });

    test('swapTiles should swap tiles and trigger events', () => {
        // 测试swapTiles方法，确保它交换瓷砖并触发事件
        // 你需要根据swapTiles方法的实现来编写具体的测试用例
    });

    test('findMatches should detect all matches correctly', () => {
        // 测试findMatches方法，确保它能正确检测所有匹配
        // 你需要根据findMatches方法的实现来编写具体的测试用例
    });

    test('removeMatches should remove matches and update the board', () => {
        // 测试removeMatches方法，确保它移除匹配并更新板上的瓷砖
        // 你需要根据removeMatches方法的实现来编写具体的测试用例
    });

    test('refill should fill empty spaces with new tiles', () => {
        // 测试refill方法，确保它填补空位并生成新的瓷砖
        // 你需要根据refill方法的实现来编写具体的测试用例
    });

    // ...添加更多测试用例
});
